# Faster multigrid solve using shared memory