This PR combines four Gauss–Seidel iterations into a single GPU kernel using shared memory.

Test with `4095*4095*1000` cells, time taken for HPMG:
```
A100:
old: 32.64 s
new: 14.91 s (+120% speedup)

MI250X:
old: 49.35 s
new: 19.87 s (+148% speedup)
```

- [ ] **Small enough** (< few 100s of lines), otherwise it should probably be split into smaller PRs
- [ ] **Tested** (describe the tests in the PR description)
- [ ] **Runs on GPU** (basic: the code compiles and run well with the new module)
- [ ] **Contains an automated test** (checksum and/or comparison with theory)
- [ ] **Documented**: all elements (classes and their members, functions, namespaces, etc.) are documented
- [ ] **Constified** (All that can be `const` is `const`)
- [ ] **Code is clean** (no unwanted comments, )
- [ ] **Style and code conventions** are respected at the bottom of https://github.com/Hi-PACE/hipace
- [ ] **Proper label and GitHub project**, if applicable